Update job engine to support cron
