+++ package.json
@@ -56,6 +56,7 @@
     "@types/marked": "^4.0.7",
     "blurhash": "^2.0.4",
     "cookies-next": "^2.1.1",
+    "cron-parser": "^4.6.0",
     "dayjs": "^1.11.6",
     "embla-carousel-react": "^7.0.3",
     "gray-matter": "^4.0.3",
