+++ src/pages/api/webhooks/run-jobs.ts
@@ -1,0 +1,45 @@
+import { WebhookEndpoint } from '~/server/common/endpoint-helpers';
+import { scanFilesJob } from '~/server/jobs/scan-files';
+import { updateMetricsJob } from '~/server/jobs/update-metrics';
+import cronParser from 'cron-parser';
+import dayjs from 'dayjs';
+
+const jobs = [scanFilesJob, updateMetricsJob];
+
+export default WebhookEndpoint(async (req, res) => {
+  const ran = [];
+
+  const now = new Date();
+  for (const { name, cron, run } of jobs) {
+    if (!isCronMatch(cron, now)) continue;
+    await run();
+    ran.push(name);
+  }
+
+  res.status(200).json({ ok: true, ran });
+});
+
+// https://github.com/harrisiirak/cron-parser/issues/153#issuecomment-590099607
+const cronScopes = ['minute', 'hour', 'day', 'month', 'weekday'] as const;
+function isCronMatch(
+  cronExpression: string,
+  date: Date,
+  scope: typeof cronScopes[number] = 'minute'
+): boolean {
+  const scopeIndex = cronScopes.indexOf(scope);
+  const day = dayjs(date);
+
+  try {
+    const { fields } = cronParser.parseExpression(cronExpression);
+
+    if (scopeIndex <= 0 && !(fields.minute as number[]).includes(day.minute())) return false;
+    if (scopeIndex <= 1 && !(fields.hour as number[]).includes(day.hour())) return false;
+    if (scopeIndex <= 2 && !(fields.dayOfMonth as number[]).includes(day.date())) return false;
+    if (scopeIndex <= 3 && !(fields.month as number[]).includes(day.month() + 1)) return false;
+    if (scopeIndex <= 4 && !(fields.dayOfWeek as number[]).includes(day.day())) return false;
+
+    return true;
+  } catch (e) {
+    return false;
+  }
+}
