+++ src/server/db/client.ts
@@ -1,7 +1,6 @@
 // src/server/db/client.ts
 import { PrismaClient } from '@prisma/client';
 import { env } from '~/env/server.mjs';
-import { fileValidationMiddleware } from './middleware/file-validation-middleware';
 
 declare global {
   // eslint-disable-next-line no-var, vars-on-top
@@ -13,8 +12,6 @@ export const prisma =
   new PrismaClient({
     log: env.NODE_ENV === 'development' ? ['query', 'error', 'warn'] : ['error'],
   });
-
-prisma.$use(fileValidationMiddleware);
 
 if (env.NODE_ENV !== 'production') {
   global.prisma = prisma;
